caa: Add option to skip Peer pod VM delete
During development, several of us have hit times where being able to
access a failing peer pod VM has been crucial to debugging issues.
The general approach to this is to add a sleep before, or comment out
`DeleteInstance`, but this involved rebuilding the CAA image, so instead
we can provide an option to set `PEERPODS_DEVELOPER_MODE=true`
in peer-pods-cm which will skip this peer pod vm delete.

+++ src/cloud-api-adaptor/pkg/adaptor/cloud/cloud.go
@@ -56,6 +56,7 @@ type ServerConfig struct {
 	PeerPodsLimitPerNode    int
 	RootVolumeSize          int
 	EnableScratchSpace      bool
+	DeveloperMode           bool
 }
 
 var logger = log.New(log.Writer(), "[adaptor/cloud] ", log.LstdFlags|log.Lmsgprefix)
@@ -475,11 +476,15 @@ func (s *cloudService) StopVM(ctx context.Context, req *pb.StopVMRequest) (*pb.S
 		sandbox.sshClientInst.DisconnectPP(string(sid))
 	}
 
-	if err := s.provider.DeleteInstance(ctx, sandbox.instanceID); err != nil {
-		logger.Printf("Error deleting an instance %s: %v", sandbox.instanceID, err)
-	} else if s.ppService != nil {
-		if err := s.ppService.ReleasePeerPod(sandbox.podName, sandbox.podNamespace, sandbox.instanceID); err != nil {
-			logger.Printf("failed to release PeerPod %v", err)
+	if s.serverConfig.DeveloperMode {
+		logger.Printf("Running in developer mode, so leaving instance %s running", sandbox.instanceID)
+	} else {
+		if err := s.provider.DeleteInstance(ctx, sandbox.instanceID); err != nil {
+			logger.Printf("Error deleting an instance %s: %v", sandbox.instanceID, err)
+		} else if s.ppService != nil {
+			if err := s.ppService.ReleasePeerPod(sandbox.podName, sandbox.podNamespace, sandbox.instanceID); err != nil {
+				logger.Printf("failed to release PeerPod %v", err)
+			}
 		}
 	}
